Standardize handling of "---" marker lines
Before, I'd never defined what happens to that opening "---".
get_yaml_text() was returning it with the rest of the YAML text.
Now, the opening "---" is always stripped and is documented as such.

get_yaml_text() should return a valid YAML document. The spec allows me
to include the opening "---" or not. I'm not sure why I chose to do that
initially, but it doesn't seem to make sense now.

As a result of that change, we always make a simple and proper
distinction between an empty section and none at all.