Objects may now only be created from strings
I was trying to be flexible and convenient. Since it's now so easy and
common to slurp whole files into a scalar (with File::Slurp,
Path::Class, and others), it's not so useful to do the I/O inside the
module anymore.

Also, this means I no longer have to worry about the encoding of the
disk file that I'm not reading. :)

Thanks to David Golden for the suggestion.