Squelch warning with empty data section on newer Perls
The warning is "Use of uninitialized value $fh in concatenation (.) or
string". Contrary to my first intepretation, $fh is not uninitialized
here; the result of <$fh> is uninitialized -- indeed, it's at EOF when
this happens.

(If it were really the filehandle $fh that were undef, the warning would
be something like "Use of uninitialized value $fh in <HANDLE>".)

This is on 5.20.2; I haven't seen this warning on older Perls.